+++ src/platform/tests/Bridge/Scaleway/Embeddings/ResultConverterTest.php
@@ -1,0 +1,60 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Bridge\Scaleway\Embeddings;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\AI\Platform\Bridge\Scaleway\Embeddings\ResultConverter;
+use Symfony\AI\Platform\Result\RawHttpResult;
+use Symfony\Contracts\HttpClient\ResponseInterface;
+
+/**
+ * @author Marcus Stöhr <marcus@fischteich.net>
+ */
+final class ResultConverterTest extends TestCase
+{
+    public function testItConvertsAResponseToAVectorResult()
+    {
+        $result = $this->createStub(ResponseInterface::class);
+        $result
+            ->method('toArray')
+            ->willReturn(json_decode($this->getEmbeddingStub(), true));
+
+        $vectorResult = (new ResultConverter())->convert(new RawHttpResult($result));
+        $convertedContent = $vectorResult->getContent();
+
+        $this->assertCount(2, $convertedContent);
+
+        $this->assertSame([0.3, 0.4, 0.4], $convertedContent[0]->getData());
+        $this->assertSame([0.0, 0.0, 0.2], $convertedContent[1]->getData());
+    }
+
+    private function getEmbeddingStub(): string
+    {
+        return <<<'JSON'
+            {
+              "object": "list",
+              "data": [
+                {
+                  "object": "embedding",
+                  "index": 0,
+                  "embedding": [0.3, 0.4, 0.4]
+                },
+                {
+                  "object": "embedding",
+                  "index": 1,
+                  "embedding": [0.0, 0.0, 0.2]
+                }
+              ]
+            }
+            JSON;
+    }
+}