+++ src/platform/tests/Bridge/Scaleway/PlatformFactoryTest.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Bridge\Scaleway;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\AI\Platform\Bridge\Scaleway\PlatformFactory;
+use Symfony\AI\Platform\Platform;
+use Symfony\Component\HttpClient\EventSourceHttpClient;
+use Symfony\Component\HttpClient\MockHttpClient;
+
+/**
+ * @author Marcus Stöhr <marcus@fischteich.net>
+ */
+final class PlatformFactoryTest extends TestCase
+{
+    public function testItCreatesPlatformWithDefaultSettings()
+    {
+        $platform = PlatformFactory::create('scaleway-test-api-key');
+
+        $this->assertInstanceOf(Platform::class, $platform);
+    }
+
+    public function testItCreatesPlatformWithCustomHttpClient()
+    {
+        $httpClient = new MockHttpClient();
+        $platform = PlatformFactory::create('scaleway-test-api-key', $httpClient);
+
+        $this->assertInstanceOf(Platform::class, $platform);
+    }
+
+    public function testItCreatesPlatformWithEventSourceHttpClient()
+    {
+        $httpClient = new EventSourceHttpClient(new MockHttpClient());
+        $platform = PlatformFactory::create('scaleway-test-api-key', $httpClient);
+
+        $this->assertInstanceOf(Platform::class, $platform);
+    }
+}